requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Pillow==12.0.0
